feature(Api):Refactor code
[Finishes #167729789]
